feat: move gold to cropXYZ
