+++ src/contextualizers/protocol/cropXYZ/abis/PackActivationSource.ts
