+++ src/contextualizers/protocol/cropXYZ/constants.ts
