+++ src/contextualizers/protocol/cropXYZ/destination.spec.ts
