+++ src/contextualizers/protocol/cropXYZ/source.spec.ts
